// src/app/administration/pages/gestor/pages/valuationQuotes/interfaces/valuation-quotes.interface.ts
// import { DataValuationQuotes } from './valuation-quotes.interface';
// Generated by https://quicktype.io

export interface ValuationQuotes {
    code:             number;
    status:           string;
    valuation_quotes: ValuationQuotesClass;
}

export interface ValuationQuotesClass {
    current_page:   number;
    data:           DataValuationQuotes[];
    first_page_url: string;
    from:           number;
    last_page:      number;
    last_page_url:  string;
    links:          Link[];
    next_page_url:  null;
    path:           string;
    per_page:       number;
    prev_page_url:  null;
    to:             number;
    total:          number;
}

export interface DataValuationQuotes {
    sycID:           number;
    id:      number | null;
    brandName:    string;
    carmodelName: string;
    userName:     string;
    surname:      string;
    email:        string;
    phone1:       number;
    subsidiary:   Subsidiary;
}

export enum Subsidiary {
    Pachuca = "pachuca",
    Puebla = "puebla",
    Tlaxcala = "tlaxcala",
}

export interface Link {
    url:    null | string;
    label:  string;
    active: boolean;
}

export interface GetActiveValuator {
    code:            number;
    status:          string;
    active_valuator: ActiveValuator[];
}

export interface ActiveValuator {
    id:      number;
    name:    string;
    surname: string;
}

export interface GetAssignValuator {
    code:    number;
    status:  string;
    message: string;
}